refactor(shared): rename Navigation to Header
