+++ pages/partner/detail/index.jsx
@@ -3,7 +3,7 @@ import { useRouter } from 'next/router';
 import { useSelector, useDispatch } from 'react-redux';
 
 import styled from '@emotion/styled';
-import Navigation from '@/shared/components/Navigation';
+import Navigation from '@/shared/components/Header';
 import Footer from '@/shared/components/Footer_v2';
 import Profile from '@/components/Profile';
 import {
