+++ pages/profile/myprofile/index.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import styled from '@emotion/styled';
 import { useAuth } from '@/contexts/Auth';
-import Navigation from '@/shared/components/Navigation';
+import Navigation from '@/shared/components/Header';
 import Footer from '@/shared/components/Footer_v2';
 import Profile from '@/components/Profile';
 import { NavigationProvider } from '@/contexts/Navigation';
